perf(codegen): PHI-based for-in loops + GEP list access + nounwind willreturn
- Convert for-in range loops from alloca-based to PHI-based induction variable
  (canonical loop form that LLVM vectorizer expects)
- Add nounwind willreturn attributes to @inline methods (enables LICM hoisting)
- Change list.get/set to use GEP (ptr_offset) for element access instead of
  inttoptr arithmetic — preserves pointer provenance for alias analysis
- Fix ptr_offset intrinsic: *Unit defaults to i8 (byte offset) not i32
- List Iteration: 4.0B → 5.4B ops/s (+35%)

+++ compiler/src/codegen/llvm/builtins/intrinsics.cpp
@@ -1106,8 +1106,8 @@ auto LLVMIRGen::try_gen_intrinsic(const std::string& fn_name, const parser::Call
             std::string ptr = gen_expr(*call.args[0]);
 
             // Infer element type from Ptr[T] or ref T
-            // Default to i32 for *Unit (void*) to match I32-sized memory operations
-            std::string elem_type = "i32";
+            // Default to i8 for *Unit (void*) so offset counts bytes, not i32-sized slots
+            std::string elem_type = "i8";
             types::TypePtr arg_type = infer_expr_type(*call.args[0]);
             if (arg_type) {
                 auto is_unit_type = [](const types::TypePtr& t) {
@@ -1119,7 +1119,7 @@ auto LLVMIRGen::try_gen_intrinsic(const std::string& fn_name, const parser::Call
 
                 if (arg_type->is<types::PtrType>()) {
                     auto inner = arg_type->as<types::PtrType>().inner;
-                    // For *Unit (void*), default to i32 (for I32 memory operations)
+                    // For *Unit (void*), keep i8 (byte-level offset)
                     if (!is_unit_type(inner)) {
                         elem_type = llvm_type_from_semantic(inner);
                     }
